started adding ranged enemy logic
- added debug labels to see health and state
- started adding hurt logic
- started adding player detection logic
- See TODO comment

+++ Resources/PlayerStats.cs
@@ -10,8 +10,8 @@ public partial class PlayerStats : Resource
     [Export] public float MaxSpeed { get; set; }
     [Export] public float Acceleration { get; set; }
     [Export] public float Friction { get; set; }
-    [Export] public float MaxHealth { get; set; }
-    [Export] public float CurrentHealth{ get; set; }
+    [Export] public int MaxHealth { get; set; }
+    [Export] public int CurrentHealth{ get; set; }
     [Export] public float JumpVelocity{ get; set; }
     [Export] public float Gravity { get; set; }
     [Export] public float WallSlideGravity { get; set; }
@@ -32,7 +32,7 @@ public partial class PlayerStats : Resource
     public PlayerState State;
     
     // Methods
-    public void TakeDamage(float damage)
+    public void TakeDamage(int damage)
     {
         CurrentHealth -= damage;
     }
